% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.bucketing.R
\name{csslr.data.bucketing}
\alias{csslr.data.bucketing}
\title{Routine to compute a bucket overview of a single variable in a data set}
\usage{
csslr.data.bucketing(
  response,
  variable,
  DT.data,
  method = "equalcut",
  DT.cut = NULL,
  numberBins = 10,
  userDefinedBins = NULL,
  bucketAverageMethod = "median",
  value.lowest = 0,
  value.highest = Inf,
  includeBinnedVariable = FALSE,
  variable.bin = NULL
)
}
\arguments{
\item{response}{Column name of the response variable contained in DT.data}

\item{variable}{Columns name of the variable that should be bucketed}

\item{DT.data}{data.table object containing both response and variable}

\item{method}{Bucketing method (only relevant for numeric variables); admissible values are "equalcut"
which would distribute the data into roughly equally sized buckets and "userdefined"
which would utilize the user-defined bucket limits for bucketing (default: "equalcut")}

\item{DT.cut}{An optional alternative data set that could be used to define buckets when the
method "equalcut" is applied. This could be useful, e.g. in cases where a mising values
imputation algorithm distorts splitting data into equally sized buckets (default: NULL)}

\item{numberBins}{Number of buckets that should be created (only relevant for numerical variables in
combination with the "equalcut" method). Default: 10}

\item{userDefinedBins}{A numeric vector containing the boundaries of buckets (only relevant for numerical
variables in combination with the "userdefined" method)}

\item{bucketAverageMethod}{In the summary table, an average value of the variable is created;
It can be controlled if this should be done using "mean" or "median"}

\item{value.lowest}{Lower boundary of the lowest bucket (only relevant for numeric variables in
combination with the "equalcut" method). Default: 0.0}

\item{value.highest}{Upper boundary of the highest bucket (only relevant for numeric variables in
combination with the "equalcut" method). Default: +Inf}

\item{includeBinnedVariable}{Flag which controls whether the bucketed variable should be included
into DT.data (Default: FALSE)}

\item{variable.bin}{Name of the bucketed variable in case it is included into the data set. If no variable
name is provided, a variable name is created automatically using the ending .bin, where
a check if performed if the variable exists in the data already (Default: NULL)}
}
\value{
A data.table containing the summary of the bucketing
}
\description{
This routine splits a variable into buckets and computes an overview table
             containing the mean response and the number of observations. For 0/1 response
             variables in addition Weight-of-Evidence / Likelihood Ratio is computed
             for each bucket. In the case of numeric
             variable observations are either split into equally sized buckets or
             user-defined bucket limits can be defined. The user has the option to include
             the bucketed variable into the data table. For categorical variables or
             numeric variables with a small number of distinct values, the buckets
             are defined by the available categories and no bucketing algorithm is needed
}
\examples{

set.seed(123)
DT.data <- data.table(Response = c(rep(0,100),rep(1,100)))
DT.data[Response == 0, Variable := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Variable := rnorm(100, -1.0, 1.0)]

bucket1 <- csslr.data.bucketing("Response", "Variable", DT.data)
bucket2 <- csslr.data.bucketing("Response", "Variable", DT.data, method = "userdefined",
                                userDefinedBins = c(-2.0,-1.0,0.0,1.0,2.0))
 
}
