% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.overview.R
\name{csslr.data.overview}
\alias{csslr.data.overview}
\title{Routine to generate an overview of the data set}
\usage{
csslr.data.overview(response, DT.data)
}
\arguments{
\item{response}{Column name of the response variable contained in DT.data}

\item{DT.data}{Data set containing the response variable and all independent variables}
}
\value{
A data.table containing all variables and their AUC and MSE values
}
\description{
The routine computes AUC and MSE for all variables in the data set;
this helps in identifying the subset of variables that should be input into the CSSLR
algorithm and the complimentary set of variables that could be excluded
}
\examples{

set.seed(123)
DT.data <- data.table(Response = c(rep(0,100),rep(1,100)))
DT.data[Response == 0, Strong1 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong1 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Strong2 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong2 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Weak1 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak1 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Weak2 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak2 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 0, Random2 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random2 := rnorm(100, 0.0, 1.0)]

DT.overview <- csslr.data.overview('Response', DT.data)

print(DT.overview)

}
