% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.analysis.lr.R
\name{csslr.model.analysis.lr}
\alias{csslr.model.analysis.lr}
\title{Routine for performing incremental analysis of a logistic regression model}
\usage{
csslr.model.analysis.lr(
  modelFormula,
  DT.data,
  panelDataIdentifier = "",
  roc = TRUE,
  calib = TRUE,
  ic = TRUE,
  lr = FALSE,
  quiet = FALSE,
  incrementVariable = TRUE,
  fullModelTrim = TRUE,
  dataValidation = TRUE
)
}
\arguments{
\item{modelFormula}{Model for which the incremental analysis should be performed}

\item{DT.data}{Data table containing all variables needed for model estimation}

\item{panelDataIdentifier}{For panel data sets a variable identifying data sets
belonging to the same entity can be defined. This leads to adjustments in the
p-values for testing the significance of model coefficients}

\item{roc}{If TRUE area under the ROC curve together with a test on difference
of AUROC between the full and the reduced models is performed}

\item{calib}{If TRUE MSE is computed together with a calibration check and the
test on difference of MSE between the full and the reduced models}

\item{ic}{If TRUE the information criteria AIC and BIC are computed for the
full and the reduced models}

\item{lr}{If TRUE likelihood ratio tests for model coefficients are performed
in addition to the Wald test that is carried out by default}

\item{quiet}{If TRUE output on the progress of the function is suppressed}

\item{incrementVariable}{If TRUE the incremental test is performed on variable
level not model part level. This means, if a variable occurs in multiple
parts of a model formula, e.g. due to interactions, it is completely
removed to evaluate its impact}

\item{fullModelTrim}{Trim the glm model results object to save memory but still
be able to do predictions with the model}

\item{dataValidation}{If true, checks for NA and Inf in the data are performed
and invalid data sets are removed; if FALSE it is assumed that input data is clean}
}
\value{
A list of tables containing the outcome of the various diagnostics
for the full and the reduced models
}
\description{
This routine performs an incremental analysis of a logistic
regression model. Each variable is removed one-by-one and the impact of each
variable on the full model is measured in terms of information criteria, area
under the ROC curve and MSE. Furthermore, variance inflation factors are
computed to allow a check for multi-collinearity
}
\examples{

set.seed(123)
DT.data <- data.table(Response = c(rep(0,100),rep(1,100)))
DT.data[Response == 0, Strong1 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong1 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Strong2 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong2 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Weak1 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak1 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Weak2 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak2 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 0, Random2 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random2 := rnorm(100, 0.0, 1.0)]

modelFormula <- Response ~ Strong1 + Strong2 + Random2
modelResults <- csslr.model.analysis.lr(modelFormula, DT.data)

# Print model summary
print(modelResults$model.summary)

# Predict probability of being "bad"
badProbs <- predict(modelResults$full.model, type = 'response', DT.data)

# Incremental Analysis
print(modelResults$ic)
print(modelResults$roc)
print(modelResults$roc.test)
print(modelResults$calib)
print(modelResults$calib.test)

}
