% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.compare.lr.R
\name{csslr.model.compare.lr}
\alias{csslr.model.compare.lr}
\title{Routine for comparing two logistic regression models}
\usage{
csslr.model.compare.lr(...)
}
\arguments{
\item{lr1}{Logistic regression Model 1 (provided either as glm object or formula)}

\item{lr2}{Logistic regression Model 2 (provided either as glm object or formula)}

\item{DT.data}{Data set containing all variables needed for model estimation}
}
\value{
Object containing the estimation results of the individual models,
information criteria and p-values of tests for difference in AUROC and MSE
}
\description{
This routine compares two logistic regression model. It computes the
mean probability forecast of each model, performs tests on difference
in AUROC and MSE using the tests in pROC and csslr.stats
}
\examples{

set.seed(123)
DT.data <- data.table(Response = c(rep(0,100),rep(1,100)))
DT.data[Response == 0, Strong1 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong1 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Strong2 := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Strong2 := rnorm(100, -1.0, 1.0)]
DT.data[Response == 0, Weak1 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak1 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Weak2 := rnorm(100, 0.5, 1.0)]
DT.data[Response == 1, Weak2 := rnorm(100, -0.5, 1.0)]
DT.data[Response == 0, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random1 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 0, Random2 := rnorm(100, 0.0, 1.0)]
DT.data[Response == 1, Random2 := rnorm(100, 0.0, 1.0)]

modelFormula1 <- Response ~ Strong1 + Strong2 + Random2
modelFormula2 <- Response ~ Strong2 + Weak1 + Weak2 + Random1

modelComparison <- csslr.model.compare.lr(modelFormula1, modelFormula2, DT.data)

# Print the outcome
print(modelComparison)

}
