% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.calib.check.R
\name{csslr.stats.calib.check}
\alias{csslr.stats.calib.check}
\title{Spiegelhalter test on calibration quality of a single probability forecast}
\usage{
csslr.stats.calib.check(indicators, probabilities)
}
\arguments{
\item{indicators}{Vector of 0/1 values where 0 is a "good" observation
and "1" models a bad observation}

\item{probabilities}{Probability forecasts of a model}
}
\value{
List with MSE, expected MSE for the probability forecast together
with test statistic and p-value of the Spiegelhalter test
}
\description{
This routine implements the test of Spiegelhalter (Spiegelhalter D. J. (1986),
Probabilistic prediction in patient management and clinical trials, Statistics
in Medicine 5(5), 421–433). The null hypothesis is that the MSE is equal to
its expected value. If the null can be rejected this is a strong indication
that a model is poorly specified
}
\examples{

set.seed(123)
DT.data <- data.table(Response = c(rep(0,100),rep(1,100)))
DT.data[Response == 0, Variable := rnorm(100, 1.0, 1.0)]
DT.data[Response == 1, Variable := rnorm(100, -1.0, 1.0)]

responseVec <- DT.data[, Response]
variableVec <- DT.data[, Variable]

calibCheck <- csslr.stats.calib.check(responseVec, variableVec)

print(calibCheck$MSE)
print(calibCheck$EMSE)
print(calibCheck$p.value)

}
