% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.genreport.R
\name{csslr.utils.genreport}
\alias{csslr.utils.genreport}
\title{Routine to create a model selection report}
\usage{
csslr.utils.genreport(
  SelectionOutput,
  DT.data,
  detailed = TRUE,
  glossary = TRUE
)
}
\arguments{
\item{SelectionOutput}{Output list generated by csslr.model.selection}

\item{DT.data}{Data set containing the response variable and all independent variables}

\item{detailed}{If TRUE all models selection steps are reported in separate tables;
if FALSE only the leading and equivalent models are reported without details on the
selection algorithm}
}
\description{
Routine to create a PDF report of an automated logistic regression model selection.
From the output of csslr.model.selection a RMarkdown document is built in a temporary folder.
This document is rendered and a PDF is written into the current working directory
}
