% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.mode.R
\name{csslr.stats.mode}
\alias{csslr.stats.mode}
\title{Function to compute the mode of a vector}
\usage{
csslr.stats.mode(vec, na.rm = FALSE)
}
\arguments{
\item{vec}{Input vector for which the mode should be computed}
}
\value{
Mode of the input vector, i.e., its most frequent value
}
\description{
Function to compute the mode of a vector
}
